# backend/openapi-schema.yaml
openapi: 3.0.3
info:
  title: E-Kalolsavam API
  version: 1.0.0
  description: OpenAPI contract for frontend-backend integration.
paths:
...
  /api/auth/users/{id}/:
    get:
      operationId: api_auth_users_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserUpdate'
          description: ''
    put:
      operationId: api_auth_users_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminUserUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminUserUpdate'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserUpdate'
          description: ''
    patch:
      operationId: api_auth_users_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedAdminUserUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedAdminUserUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedAdminUserUpdate'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserUpdate'
          description: ''
    delete:
      operationId: api_auth_users_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
  /api/token/:
    post:
      operationId: api_token_create
      description: |-
        Takes a set of user credentials and returns an access and refresh JSON web
        token pair to prove the authentication of those credentials.
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenObtainPair'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenObtainPair'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TokenObtainPair'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenObtainPair'
          description: ''
...
components:
  schemas:
...
    AdminUserUpdate:
      type: object
      properties:
        role:
          $ref: '#/components/schemas/Role083Enum'
        phone:
          type: string
          maxLength: 15
        approval_status:
          $ref: '#/components/schemas/ApprovalStatusEnum'
        student_class:
          type: integer
          maximum: 65535
          minimum: 0
          nullable: true
        school:
          type: integer
          nullable: true
        school_category_extra:
          type: string
          nullable: true
          maxLength: 10
        gender:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/Gender8c3Enum'
          - $ref: '#/components/schemas/BlankEnum'
          - $ref: '#/components/schemas/NullEnum'
      required:
      - role
...
    ApprovalStatusEnum:
      enum:
      - pending
      - approved
      - rejected
      type: string
      description: |-
        * `pending` - Pending
        * `approved` - Approved
        * `rejected` - Rejected
...
    BlankEnum:
      enum:
      - ''
...
    Gender8c3Enum:
      enum:
      - BOYS
      - GIRLS
      type: string
      description: |-
        * `BOYS` - Boys
        * `GIRLS` - Girls
...
    NullEnum:
      enum:
      - null
...
    PatchedAdminUserUpdate:
      type: object
      properties:
        role:
          $ref: '#/components/schemas/Role083Enum'
        phone:
          type: string
          maxLength: 15
        approval_status:
          $ref: '#/components/schemas/ApprovalStatusEnum'
        student_class:
          type: integer
          maximum: 65535
          minimum: 0
          nullable: true
        school:
          type: integer
          nullable: true
        school_category_extra:
          type: string
          nullable: true
          maxLength: 10
        gender:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/Gender8c3Enum'
          - $ref: '#/components/schemas/BlankEnum'
          - $ref: '#/components/schemas/NullEnum'
...
    Role083Enum:
      enum:
      - student
      - judge
      - admin
      - volunteer
      - school
      type: string
      description: |-
        * `student` - Student
        * `judge` - Judge
        * `admin` - Admin
        * `volunteer` - Volunteer
        * `school` - School
...
    TokenObtainPair:
      type: object
      properties:
        username:
          type: string
          writeOnly: true
        password:
          type: string
          writeOnly: true
        access:
          type: string
          readOnly: true
        refresh:
          type: string
          readOnly: true
      required:
      - access
      - password
      - refresh
      - username
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT